# core-service/src/main/java/com/geekbrains/coreservice/model/Category.java
package com.geekbrains.coreservice.model;

import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;
import java.util.*;


@Data
@NoArgsConstructor
@Entity
@Table(name = "categories")
public class Category {

  @Id
  @GeneratedValue(strategy = GenerationType.IDENTITY)
  @Column(name = "id")
  private Long id;

  @Column(name = "title")
  private String title;

  @OneToMany(mappedBy = "category")
  private List<Product> products;


}